Use latest core + add on demand rules iOS
