--- nym-vpn-apple/Services/Sources/Services/Tunnels/OnDemand/OnDemandRule.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public enum OnDemandRule {
-    case off
-    case onWifi(OnDemandSSID)
-    case onWifiAndCellular(OnDemandSSID)
-    // case onWifiAndEthernet
-}
