--- nym-vpn-apple/Services/Sources/Services/Tunnels/OnDemand/OnDemandSSID.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-public enum OnDemandSSID {
-    case any
-    case onlySpecific([String])
-    case onWifiAndCellular([String])
-    //    case onWifiAndEthernet
-}
